+++ lib/SILOptimizer/Transforms/PerformanceInliner.cpp
@@ -39,14 +39,6 @@ llvm::cl::opt<bool> EnableSILInliningOfGenerics(
 
 namespace {
 
-// Controls the decision to inline functions with @_semantics, @effect and
-// global_init attributes.
-enum class InlineSelection {
-  Everything,
-  NoGlobalInit, // and no availability semantics calls
-  NoSemanticsAndGlobalInit
-};
-
 using Weight = ShortestPathAnalysis::Weight;
 
 class SILPerformanceInliner {
@@ -132,8 +124,6 @@ class SILPerformanceInliner {
     return SPA;
   }
 
-  SILFunction *getEligibleFunction(FullApplySite AI);
-
   bool isProfitableToInline(FullApplySite AI,
                             Weight CallerWeight,
                             ConstantTracker &callerTracker,
@@ -161,160 +151,6 @@ public:
 };
 
 } // end anonymous namespace
-
-// Return true if the callee has self-recursive calls.
-static bool calleeIsSelfRecursive(SILFunction *Callee) {
-  for (auto &BB : *Callee)
-    for (auto &I : BB)
-      if (auto Apply = FullApplySite::isa(&I))
-        if (Apply.getReferencedFunction() == Callee)
-          return true;
-  return false;
-}
-
-// Returns true if the callee contains a partial apply instruction,
-// whose substitutions list would contain opened existentials after
-// inlining.
-static bool calleeHasPartialApplyWithOpenedExistentials(FullApplySite AI) {
-  if (!AI.hasSubstitutions())
-    return false;
-
-  SILFunction *Callee = AI.getReferencedFunction();
-  auto Subs = AI.getSubstitutions();
-
-  // Bail if there are no open existentials in the list of substitutions.
-  bool HasNoOpenedExistentials = true;
-  for (auto Sub : Subs) {
-    if (Sub.getReplacement()->hasOpenedExistential()) {
-      HasNoOpenedExistentials = false;
-      break;
-    }
-  }
-
-  if (HasNoOpenedExistentials)
-    return false;
-
-  auto SubsMap = Callee->getLoweredFunctionType()
-    ->getGenericSignature()->getSubstitutionMap(Subs);
-
-  for (auto &BB : *Callee) {
-    for (auto &I : BB) {
-      if (auto PAI = dyn_cast<PartialApplyInst>(&I)) {
-        auto PAISubs = PAI->getSubstitutions();
-        if (PAISubs.empty())
-          continue;
-
-        // Check if any of substitutions would contain open existentials
-        // after inlining.
-        auto PAISubMap = PAI->getOrigCalleeType()
-          ->getGenericSignature()->getSubstitutionMap(PAISubs);
-        PAISubMap = PAISubMap.subst(SubsMap);
-        if (PAISubMap.hasOpenedExistential())
-          return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-// Returns the callee of an apply_inst if it is basically inlineable.
-SILFunction *SILPerformanceInliner::getEligibleFunction(FullApplySite AI) {
-
-  SILFunction *Callee = AI.getReferencedFunction();
-
-  if (!Callee) {
-    return nullptr;
-  }
-
-  // Don't inline functions that are marked with the @_semantics or @effects
-  // attribute if the inliner is asked not to inline them.
-  if (Callee->hasSemanticsAttrs() || Callee->hasEffectsKind()) {
-    if (WhatToInline == InlineSelection::NoSemanticsAndGlobalInit) {
-      ArraySemanticsCall ASC(AI.getInstruction());
-      if (!ASC.canInlineEarly())
-        return nullptr;
-    }
-    // The "availability" semantics attribute is treated like global-init.
-    if (Callee->hasSemanticsAttrs() &&
-        WhatToInline != InlineSelection::Everything &&
-        Callee->hasSemanticsAttrThatStartsWith("availability")) {
-      return nullptr;
-    }
-  } else if (Callee->isGlobalInit()) {
-    if (WhatToInline != InlineSelection::Everything) {
-      return nullptr;
-    }
-  }
-
-  // We can't inline external declarations.
-  if (Callee->empty() || Callee->isExternalDeclaration()) {
-    return nullptr;
-  }
-
-  // Explicitly disabled inlining.
-  if (Callee->getInlineStrategy() == NoInline) {
-    return nullptr;
-  }
-  
-  if (!Callee->shouldOptimize()) {
-    return nullptr;
-  }
-
-  SILFunction *Caller = AI.getFunction();
-
-  // We don't support inlining a function that binds dynamic self because we
-  // have no mechanism to preserve the original function's local self metadata.
-  if (mayBindDynamicSelf(Callee)) {
-    // Check if passed Self is the same as the Self of the caller.
-    // In this case, it is safe to inline because both functions
-    // use the same Self.
-    if (AI.hasSelfArgument() && Caller->hasSelfParam()) {
-      auto CalleeSelf = stripCasts(AI.getSelfArgument());
-      auto CallerSelf = Caller->getSelfArgument();
-      if (CalleeSelf != SILValue(CallerSelf))
-        return nullptr;
-    } else
-      return nullptr;
-  }
-
-  // Detect self-recursive calls.
-  if (Caller == Callee) {
-    return nullptr;
-  }
-
-  // A non-fragile function may not be inlined into a fragile function.
-  if (Caller->isSerialized() &&
-      !Callee->hasValidLinkageForFragileInline()) {
-    if (!Callee->hasValidLinkageForFragileRef()) {
-      llvm::errs() << "caller: " << Caller->getName() << "\n";
-      llvm::errs() << "callee: " << Callee->getName() << "\n";
-      llvm_unreachable("Should never be inlining a resilient function into "
-                       "a fragile function");
-    }
-    return nullptr;
-  }
-
-  // Inlining self-recursive functions into other functions can result
-  // in excessive code duplication since we run the inliner multiple
-  // times in our pipeline
-  if (calleeIsSelfRecursive(Callee)) {
-    return nullptr;
-  }
-
-  if (!EnableSILInliningOfGenerics && AI.hasSubstitutions()) {
-    // Inlining of generics is not allowed.
-    return nullptr;
-  }
-
-  // IRGen cannot handle partial_applies containing opened_existentials
-  // in its substitutions list.
-  if (calleeHasPartialApplyWithOpenedExistentials(AI)) {
-    return nullptr;
-  }
-
-  return Callee;
-}
 
 // Returns true if it is possible to perform a generic
 // specialization for a given call.
@@ -650,7 +486,7 @@ void SILPerformanceInliner::collectAppliesToInline(
     // At this occasion we record additional weight increases.
     addWeightCorrection(FAS, WeightCorrections);
 
-    if (SILFunction *Callee = getEligibleFunction(FAS)) {
+    if (SILFunction *Callee = getEligibleFunction(FAS, WhatToInline)) {
       // Compute the shortest-path analysis for the callee.
       SILLoopInfo *CalleeLI = LA->get(Callee);
       ShortestPathAnalysis *CalleeSPA = getSPA(Callee, CalleeLI);
@@ -696,7 +532,7 @@ void SILPerformanceInliner::collectAppliesToInline(
 
       FullApplySite AI = FullApplySite(&*I);
 
-      auto *Callee = getEligibleFunction(AI);
+      auto *Callee = getEligibleFunction(AI, WhatToInline);
       if (Callee) {
         if (!BlockWeight.isValid())
           BlockWeight = SPA->getWeight(block, Weight(0, 0));
@@ -813,7 +649,7 @@ void SILPerformanceInliner::visitColdBlocks(
       if (!AI)
         continue;
 
-      auto *Callee = getEligibleFunction(AI);
+      auto *Callee = getEligibleFunction(AI, WhatToInline);
       if (Callee && decideInColdBlock(AI, Callee)) {
         AppliesToInline.push_back(AI);
       }
